requirements/base.txt
amqp==2.5.2
asgiref==3.2.7
billiard==3.6.3.0
celery==4.4.2
Django==3.0.6
django-environ==0.4.5
djangorestframework==3.11.0
importlib-metadata==1.6.0
kombu==4.6.8
Pillow==7.1.2
psycopg2==2.8.5
pytz==2020.1
redis==3.5.2
reportlab==3.5.42
sqlparse==0.3.1
vine==1.3.0
zipp==3.1.0
